# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/instance": {
            "get": {
                "description": "Get Instance.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "Instance"
                ],
                "summary": "Get Instance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GetInstanceResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResp"
                        }
                    }
                }
            },
            "put": {
                "description": "Update Instance.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "Instance"
                ],
                "summary": "Update Instance",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Instance"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.CreateInstanceResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResp"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Instance.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "Instance"
                ],
                "summary": "Create Instance",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Instance"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.CreateInstanceResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResp"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Instance.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "Instance"
                ],
                "summary": "Delete Instance",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Instance"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.CreateInstanceResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "constant.JobState": {
            "type": "string",
            "enum": [
                "Creating",
                "Updating",
                "Deleting",
                "Processing",
                "Succeed",
                "Deleted",
                "Error",
                "Unknown",
                "Timeout",
                "Finished"
            ],
            "x-enum-varnames": [
                "JobStateCreating",
                "JobStateUpdating",
                "JobStateDeleting",
                "JobStateProcessing",
                "JobStateSucceed",
                "JobStateDeleted",
                "JobStateError",
                "JobStateUnknown",
                "JobStateTimeout",
                "JobStateFinished"
            ]
        },
        "types.CreateInstanceResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "properties": {
                        "jobId": {
                            "type": "string"
                        }
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "types.ErrorResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "properties": {
                        "error": {
                            "type": "string"
                        }
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "types.GetInstanceResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/types.Instance"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "types.Instance": {
            "type": "object",
            "required": [
                "instanceName",
                "namespace"
            ],
            "properties": {
                "instanceName": {
                    "type": "string"
                },
                "jobID": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "spec": {
                    "description": "Specification of the desired behavior of a Instance.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.InstanceSpec"
                        }
                    ]
                },
                "status": {
                    "description": "Current status of a Instance",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.InstanceStatus"
                        }
                    ]
                }
            }
        },
        "types.InstanceSpec": {
            "type": "object",
            "required": [
                "image",
                "servers"
            ],
            "properties": {
                "hostID": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "keepStorage": {
                    "type": "boolean"
                },
                "replicas": {
                    "type": "integer"
                },
                "servers": {
                    "type": "array",
                    "maxItems": 30,
                    "minItems": 1,
                    "uniqueItems": true,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "types.InstanceStatus": {
            "type": "object",
            "properties": {
                "createdTime": {
                    "type": "string"
                },
                "jobState": {
                    "$ref": "#/definitions/constant.JobState"
                },
                "modifyTime": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    }
}